pyproject.toml
[project]
name = "localmultirag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=5.17.1",
    "langchain-chroma>=0.2.2",
    "langchain-community>=0.3.18",
    "python-dotenv>=1.0.1",
]
